// src/clients/generated/commonApi/models/createMachineRequest.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * order-manager
 * API order-manager
 * OpenAPI spec version: 1.0.0
 */
import type { CreateMachineRequestMachineType } from './createMachineRequestMachineType';

/**
 * CreateMachineRequest
 */
export interface CreateMachineRequest {
  /** Тип станка с ЧПУ. */
  machineType: CreateMachineRequestMachineType;
  /**
   * Наименование станка с ЧПУ.
   * @maxLength 255
   */
  name: string;
}