// src/clients/generated/commonApi/models/createMachineRequestMachineType.ts
/**
 * Generated by orval v7.3.0 🍺
 * Do not edit manually.
 * order-manager
 * API order-manager
 * OpenAPI spec version: 1.0.0
 */

/**
 * Тип станка с ЧПУ.
 */
export type CreateMachineRequestMachineType = typeof CreateMachineRequestMachineType[keyof typeof CreateMachineRequestMachineType];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const CreateMachineRequestMachineType = {
  THREE_AXIS: 'THREE_AXIS',
  FIVE_AXIS: 'FIVE_AXIS',
} as const;